fixed some linux compilation issues
git-svn-id: https://assimp.svn.sourceforge.net/svnroot/assimp/trunk@257 67173fc5-114c-0410-ac8e-9d2fd5bffc1f

+++ code/ColladaLoader.cpp
@@ -239,7 +239,7 @@ void ColladaLoader::BuildMeshesForNode( const ColladaParser& pParser, const Coll
 	pTarget->mNumMeshes = newMeshRefs.size();
 	if( newMeshRefs.size())
 	{
-		pTarget->mMeshes = new size_t[pTarget->mNumMeshes];
+		pTarget->mMeshes = new unsigned int[pTarget->mNumMeshes];
 		std::copy( newMeshRefs.begin(), newMeshRefs.end(), pTarget->mMeshes);
 	}
 }
